Added `GET /projects/{id}/outcome-tracker/drift`: a pure-Python tracking-drift early-warning feat...
Mode: ADD

Added `GET /projects/{id}/outcome-tracker/drift`: a pure-Python tracking-drift early-warning feature that rebuilds the production trajectory projection at each checkpoint, compares it to the next logged conversion, and reports ON_TRACK/AHEAD/BEHIND status, widening/narrowing drift, severity (ok/watc

Pass log: run-20260809-163520-attempt1.log

+++ backend/app/simulation/outcome_tracker_forecast.py
@@ -222,7 +222,7 @@ def _ceiling(target: float | None, max_observed: float) -> float:
 def _projected(
     latest: float,
     slope: float,
-    horizon_days: int,
+    horizon_days: float,
     ceiling: float,
 ) -> float:
     raw = latest + slope * horizon_days
